parse stringified array; display tasks completed
Task duration objects now successfully stringified and stored in
localStorage, then retrieved and parsed in order to build up a table of
completed tasks.

The retreived array is reversed so that the most recent task is displayed
first.

A (document).ready function ensures that the table is built on refresh or
page load, and the populate_stats_table function only attempts to build
the table if it is not null (otherwise, it passes 'No tasks completed').

The moment.js library added to provide a simple way to convert the
JavaScript Date() object into an appropriate format.

***TO DO***
Need to ensure that the state of the stop/start buttons persist through
browser close/refresh.

Also need to provide a check so that the start button can't perform its
click function unless both a Project and a Task have been selected. These
radio buttons also need their state saved.

I would like to add further statistics (i.e. total time spent on task x) -
but this might become quite fiddly with a bunch of selectors that might
clutter up the interface if I'm not careful.

Work on UX/css; ensure that the page is responsive/mobile-first.