log total time spent on each task.
New table prepared which will log the total time spent on each task.

First a function finds unique tasks and appends the to an array, then a
second function takes these and uses them as a guide for a nested
iteration over all tasks calculating the total duration of each.

The task, project, and duration is then appended to a new array.

This array is then sent to html. Unfortunatly, still in order of latest
first.

***TO DO***
Attemps are being made to ensure that the aggregate_table shows the
greatest ammount of time spent first (this seems a logical way to order
the data).

It is, however, being extremely fighty! I'm sure this should be terribly
trivial ~ I can access the values but just can't create a new array ordered
by duration.

Also, worringly, just noticed that as I have passed midnight (from the 4th
feb to the 5th), my duration from the 4th is now displaying as null! Is
this an edgecase I hadn't thought of, or something else ...

Still need to pretify.